assignment2/requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-tavily>=0.2.17
langchain-community>=0.3.0
google-genai>=1.0.0
tavily-python>=0.5.0
fastapi>=0.115.0
uvicorn>=0.32.0
streamlit>=1.40.0
requests>=2.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
